Fix `ansible-playbook -C` by allowing some tests in check mode
There are expressions which reference variables defined by tasks that
didn't execute in check mode, causing errors.  Since the tasks that
register the required variables do not change the remote state, they
have been flagged with `check_mode: no` to allow them to run when the
`-C` flag is in use.